tools/ml/requirements.txt
google-cloud-firestore
tensorflow>=2.12
numpy
scikit-learn
